+++ spec/components/checks/active/xss_dom_script_context_spec.rb
@@ -5,7 +5,7 @@ describe name_from_filename do
 
     def self.elements
         [ Element::Form::DOM, Element::Link::DOM, Element::Cookie::DOM,
-          Element::LinkTemplate::DOM, Element::Input::DOM, Element::UIForm::DOM ]
+          Element::LinkTemplate::DOM, Element::UIInput::DOM, Element::UIForm::DOM ]
     end
 
     def issue_count_per_element
@@ -14,7 +14,7 @@ describe name_from_filename do
             Element::Link::DOM         => 2,
             Element::Cookie::DOM       => 2,
             Element::LinkTemplate::DOM => 2,
-            Element::Input::DOM        => 2,
+            Element::UIInput::DOM        => 2,
             Element::UIForm::DOM       => 2
         }
     end
@@ -32,7 +32,7 @@ describe name_from_filename do
 
                 expect(data_flow_sinks.size).to eq 2
 
-            elsif issue.vector.class == Element::Input::DOM
+            elsif issue.vector.class == Element::UIInput::DOM
 
                 expect(data_flow_sinks.size).to eq 3
 
@@ -69,7 +69,7 @@ describe name_from_filename do
                     expect(trace.size).to eq 2
                     expect(trace.first.url).to eq issue.page.dom.url
 
-                when Element::Input::DOM
+                when Element::UIInput::DOM
                     transition = issue.page.dom.transitions.last
 
                     expect(transition.element.tag_name).to eq :input