+++ spec/support/factories/element/ui_input.rb
@@ -1,4 +1,4 @@
-Factory.define :input, class: Arachni::Element::Input,
+Factory.define :ui_input, class: Arachni::Element::UIInput,
                options: {
                    url:    'http://test.com',
                    source: '<input id="myname" />',
@@ -8,6 +8,6 @@ Factory.define :input, class: Arachni::Element::Input,
                    }
                }
 
-Factory.define :input_dom do
-    Factory[:input].dom
+Factory.define :ui_input_dom do
+    Factory[:ui_input].dom
 end