+++ ui/cli/framework.rb
@@ -456,15 +456,15 @@ class Framework
         if !options.audit.links? && !options.audit.forms? &&
             !options.audit.cookies? && !options.audit.headers? &&
             !options.audit.link_templates? && !options.audit.jsons? &&
-            !options.audit.xmls? && !options.audit.inputs? &&
+            !options.audit.xmls? && !options.audit.ui_inputs? &&
             !options.audit.ui_forms?
 
             print_info 'No element audit options were specified, will audit ' <<
-                           'links, forms, cookies, inputs, UI forms, JSONs and XMLs.'
+                           'links, forms, cookies, UI inputs, UI forms, JSONs and XMLs.'
             print_line
 
-            options.audit.elements :links, :forms, :cookies, :inputs, :ui_forms,
-                                   :jsons, :xmls
+            options.audit.elements :links, :forms, :cookies, :ui_inputs,
+                                   :ui_forms, :jsons, :xmls
         end
     end